+++ packages/atomic/cypress/integration/result-templates/field-condition.cypress.ts
@@ -1,0 +1,25 @@
+describe('Field Condition Component', () => {
+  describe('when not a child of an "template" element', () => {
+    it.skip('should render an "atomic-component-error" component');
+  });
+
+  it.skip('the "if-defined" prop should add a condition to the component');
+
+  it.skip('the "if-not-defined" prop should add a condition to the component');
+
+  it.skip('the "must-match-x" prop should add a condition to the component');
+
+  it.skip(
+    'the "must-not-match-x" prop should add a condition to the component'
+  );
+
+  it.skip('the "conditions" prop should add a condition(s) to the component');
+
+  describe('when all the conditions are fulfilled', () => {
+    it.skip('should render the component');
+  });
+
+  describe('when any of the conditions is unfulfilled', () => {
+    it.skip('should remove the component from the DOM');
+  });
+});