+++ packages/atomic/stencil.config.ts
@@ -8,6 +8,23 @@ import {inlineSvg} from 'stencil-inline-svg';
 import tailwind from 'tailwindcss';
 import atImport from 'postcss-import';
 import autoprefixer from 'autoprefixer';
+import replacePlugin from '@rollup/plugin-replace';
+import {readFileSync} from 'fs';
+
+const isProduction = process.env.BUILD === 'production';
+
+function getPackageVersion(): string {
+  return JSON.parse(readFileSync('package.json', 'utf-8')).version;
+}
+
+function replace() {
+  const env = isProduction ? 'production' : 'development';
+  const version = getPackageVersion();
+  return replacePlugin({
+    'process.env.NODE_ENV': JSON.stringify(env),
+    'process.env.VERSION': JSON.stringify(version),
+  });
+}
 
 const isDevWatch: boolean =
   process.argv &&
@@ -41,13 +58,14 @@ export const config: Config = {
     transformIgnorePatterns: [],
   },
   devServer: {
-    reloadStrategy: 'hmr',
+    reloadStrategy: 'pageReload',
   },
   plugins: [
     inlineSvg(),
     postcss({
       plugins: [atImport(), tailwind(), autoprefixer()],
     }),
+    replace(),
   ],
   rollupPlugins: {
     before: [