+++ packages/headless/doc-parser/src/entity-builder.test.ts
@@ -18,6 +18,23 @@ describe('#buildEntity', () => {
     expect(entity.desc).toBe('The number of pages to display in the pager.');
   });
 
+  it(`when a docComment contains an @defaultValue tag,
+  it sets the defaultValue to the default`, () => {
+    const parser = new TSDocParser();
+    const {docComment} = parser.parseString(
+      '/**\n * The number of pages to display in the pager.\n *\n * @defaultValue `5`\n */\n'
+    );
+
+    const entity = buildEntity({
+      name: 'numberOfPages',
+      type: 'number',
+      isOptional: true,
+      comment: docComment,
+    });
+
+    expect(entity.defaultValue).toBe('`5`');
+  });
+
   it(`when a docComment contains a summary and a non-standard tag (e.g. @default),
   it ignores the non-standard tag`, () => {
     const parser = new TSDocParser();