+++ packages/headless/src/app/search-app-reducers.ts
@@ -26,6 +26,7 @@ import {advancedSearchQueriesReducer} from '../features/advanced-search-queries/
 import {debugReducer} from '../features/debug/debug-slice';
 import {facetOrderReducer} from '../features/facets/facet-order/facet-order-slice';
 import {redo, snapshot, undo} from '../features/history/history-actions';
+import {versionReducer} from '../features/debug/version-slice';
 
 /**
  * Map of reducers that make up the SearchAppState.
@@ -62,4 +63,5 @@ export const searchAppReducers: ReducersMapObject<SearchAppState> = {
   pipeline: pipelineReducer,
   searchHub: searchHubReducer,
   debug: debugReducer,
+  version: versionReducer,
 };