+++ packages/headless/src/controllers/product-recommendations/headless-cart-recommendations-options.ts
@@ -6,7 +6,11 @@ export interface CartRecommendationsListOptions {
   /**
    * The maximum number of recommendations, from 1 to 50.
    *
-   * @default 5
+   * @defaultValue `5`
    */
   maxNumberOfRecommendations?: number;
+  /**
+   * Additional fields to fetch in the results.
+   */
+  additionalFields?: string[];
 }