+++ packages/quantic/force-app/main/default/lwc/searchBox/searchBox.js
@@ -1,105 +1,49 @@
-import {LightningElement, api, track} from 'lwc';
-import TributePath from '@salesforce/resourceUrl/tributejs';
+import { LightningElement, api, track } from 'lwc';
 // @ts-ignore
-import {loadScript} from 'lightning/platformResourceLoader';
-import {initializeComponent} from 'c/initialization';
+import { registerComponentForInit, initializeWithHeadless } from 'c/headlessLoader';
+
+const ENTER = 13;
+const ARROWUP = 38;
+const ARROWDOWN = 40;
 
 export default class SearchBox extends LightningElement {
   /** @type {import("coveo").SearchBoxState} */
   @track state = {
-    // @ts-ignore TODO: Check SearchBoxState typing and integration with LWC/Quantic
+    // @ts-ignore
     redirectTo: '',
     suggestions: [],
     value: '',
   };
+
   /** @type {any} */
-  @track suggestions = [];
+  get suggestions() {
+    return this.state.suggestions.map((s, index) => ({ key: index, value: s.rawValue }));
+  }
+
+  /** @type {string} */
+  @api engineId;
+  /** @type {string} */
+  @api placeholder = 'Search';
 
   /** @type {import("coveo").SearchBox} */
   searchBox;
   /** @type {import("coveo").Unsubscribe} */
   unsubscribe;
-  tributeLoaded = false;
-  /** @type {any} */
-  tribute;
+  /** @type {number} */
+  selectionIndex = -1;
+  /** @type {HTMLInputElement} */
+  input;
+  /** @type {import("lwc").HTMLElementTheGoodPart} */
+  combobox;
 
-  connectedCallback() {
-    initializeComponent(this);
-
-    if (this.tributeLoaded) {
-      return;
-    }
-
-    loadScript(this, TributePath + '/tribute.js').then(
-      () => (this.tributeLoaded = true)
-    );
-  }
-
-  disconnectedCallback() {
-    if (this.unsubscribe) {
-      this.unsubscribe();
-    }
-  }
-
-  renderedCallback() {
-    // @ts-ignore
-    const input = this.template.querySelector('input');
-    // @ts-ignore
-    const wrapper = this.template.querySelector('.slds-dropdown');
-    // @ts-ignore
-    const combobox = this.template.querySelector('.slds-combobox');
-
-    if (!input || !wrapper || !this.tributeLoaded || this.tribute) {
-      return;
-    }
-
-    this.configureTributeJS(input, wrapper, combobox);
-  }
-
-  /**
-   * @param {import("lwc").HTMLElementTheGoodPart} input
-   * @param {import("lwc").HTMLElementTheGoodPart} wrapper
-   * @param {import("lwc").HTMLElementTheGoodPart} combobox
-   */
-  configureTributeJS(input, wrapper, combobox) {
-    const tributeOptions = {
-      values: [],
-      searchOpts: {
-        skip: true,
-      },
-      selectTemplate: (item) => item.string,
-      menuContainer: wrapper,
-      positionMenu: false,
-      autocompleteMode: true,
-      replaceTextSuffix: '',
-      containerClass: 'slds-listbox slds-listbox_vertical',
-      itemClass:
-        'slds-listbox__item slds-media slds-listbox__option slds-listbox__option_plain slds-media_small',
-      noMatchTemplate: '',
-    };
-    // @ts-ignore
-    this.tribute = new Tribute(tributeOptions);
-    this.tribute.attach(input);
-
-    input.addEventListener(
-      'tribute-replaced',
-      /**
-       * @param {CustomEvent} e
-       */
-      (e) => {
-        this.searchBox.updateText(e.detail.item.string);
-        this.searchBox.submit();
-      }
-    );
-
-    input.addEventListener('tribute-active-true', () => {
-      combobox.classList.add('slds-is-open');
-    });
-
-    input.addEventListener('tribute-active-false', () => {
-      combobox.classList.remove('slds-is-open');
-    });
-  }
+  /** @type {() => void} */
+  resetSelectionIndex = () => {this.selectionIndex = -1};
+  /** @type {() => boolean} */
+  // @ts-ignore
+  areSuggestionsShown = () => (this.template.querySelector('.slds-combobox').classList.contains('slds-is-open'));
+  /** @type {() => HTMLElement[]} */
+  // @ts-ignore
+  getSuggestionElements = () => this.template.querySelectorAll('.slds-listbox__option');
 
   /**
    * @param {import("coveo").Engine} engine
@@ -110,39 +54,133 @@ export default class SearchBox extends LightningElement {
     this.unsubscribe = this.searchBox.subscribe(() => this.updateState());
   }
 
+  connectedCallback() {
+    registerComponentForInit(this, this.engineId);
+  }
+
+  renderedCallback() {
+    initializeWithHeadless(this, this.engineId, this.initialize.bind(this));
+    if (!this.input) {
+      // @ts-ignore
+      this.input = this.template.querySelector('input');
+    }
+    if (!this.combobox) {
+      // @ts-ignore
+      this.combobox = this.template.querySelector('.slds-combobox');
+    }
+  }
+
+  disconnectedCallback() {
+    if (this.unsubscribe) {
+      this.unsubscribe();
+    }
+  }
+
+  updateState() {
+    this.state = this.searchBox.state;
+  }
+
+  showSuggestions() {
+    this.combobox.classList.add('slds-is-open');
+    this.combobox.setAttribute("aria-expanded", true);
+  }
+
+  hideSuggestions() {
+    this.combobox.classList.remove('slds-is-open');
+    this.combobox.setAttribute('aria-expanded', false);
+    this.resetHighlighted();
+    this.resetSelectionIndex();
+  }
+
+  setHighlighted() {
+    const options = this.getSuggestionElements();
+    const option = options[this.selectionIndex];
+
+    this.resetHighlighted();
+    option.setAttribute('aria-selected', 'true');
+    option.classList.add('slds-has-focus');
+    this.input.value = option.textContent;
+  }
+
+  resetHighlighted() {
+    const options = this.getSuggestionElements();
+
+    options.forEach((element) => {
+      element.setAttribute('aria-selected', 'false');
+      element.classList.remove('slds-has-focus');
+    });
+  }
+
   /**
-   * @param {InputEvent & {target: {value : string}}} event
+   * @param {string} textValue
    */
-  onChange(event) {
-    this.searchBox.updateText(event.target.value);
+  updateSearchboxText(textValue) {
+    this.input.value = textValue;
+    this.searchBox.updateText(textValue);
+  }
+
+  handleEnter() {
+    if (this.selectionIndex >= 0) {
+      this.searchBox.updateText(this.input.value);
+    }
+    this.searchBox.submit();
+    this.input.blur();
+  }
+
+  handleArrowUp() {
+    this.selectionIndex--;
+    if (this.selectionIndex < 0) {
+      this.selectionIndex = this.suggestions.length - 1;
+    }
+    this.setHighlighted();
+  }
+
+  handleArrowDown() {
+    this.selectionIndex++;
+    if (this.selectionIndex > this.suggestions.length - 1) {
+      this.selectionIndex = 0;
+    }
+    this.setHighlighted();
   }
 
   /**
    * @param {KeyboardEvent & {target: {value : string}}} event
    */
   onKeyup(event) {
-    if (event.which === 13) {
-      this.searchBox.submit();
+    if (event.which === ENTER) {
+      this.handleEnter();
+    } else if (this.areSuggestionsShown() && event.which === ARROWUP) {
+      this.handleArrowUp();
+    } else if (this.areSuggestionsShown() && event.which === ARROWDOWN) {
+      this.handleArrowDown();
+    } else {
+      this.resetSelectionIndex();
+      this.resetHighlighted();
+      this.searchBox.updateText(event.target.value);
     }
-    this.searchBox.updateText(event.target.value);
-  }
-
-  updateState() {
-    this.state = this.searchBox.state;
-
-    this.suggestions = this.state.suggestions.map((s) => s.rawValue);
-    if (!this.tribute) {
-      return;
-    }
-
-    this.tribute.append(
-      0,
-      this.state.suggestions.map((s) => ({key: s.rawValue, value: s.rawValue})),
-      true
-    );
   }
 
   onFocus() {
     this.searchBox.showSuggestions();
+    this.showSuggestions();
+  }
+
+  onBlur() {
+    this.hideSuggestions();
+  }
+
+  preventDefault(event) {
+    event.preventDefault();
+  }
+
+  /**
+   * @param {KeyboardEvent & {target: {textContent : string}}} event
+   */
+  handleSuggestionSelection(event) {
+    const textValue = event.target.textContent;
+
+    this.updateSearchboxText(textValue);
+    this.searchBox.submit();
+    this.input.blur();
   }
 }